Item component extracted from Info to stand-alone file
