+++ client/src/components/shared/Info/Info.scss
@@ -47,16 +47,6 @@
   cursor: pointer;
 }
 
-.info-item {
-  padding: .5em 0;
-}
-
-.info-item-label {
-  color: $secondary;
-  font-weight: normal;
-  margin-right: .5em;
-}
-
 .info-list {
   list-style: none;
   padding: 0;
