+++ client/src/components/shared/Info/Item.scss
@@ -1,0 +1,11 @@
+@import '../../variables.scss';
+
+.info-item {
+  padding: .5em 0;
+}
+
+.info-item-label {
+  color: $secondary;
+  font-weight: normal;
+  margin-right: .5em;
+}
